requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==4.2.19
django-environ==0.12.0
pillow==11.1.0
pip==25.0.1
qrcode==8.0
setuptools==75.8.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
wheel==0.45.1